Use intentions to allow the user to perform context-sensitive actions.
This replaces cumbersome menu items that don't always apply in all situations. Intentions allow us to display actions based on the context of the action (e.g. docblocks can only be generated when the cursor is on a method name, code can only be extracted if there is a selection, etc.).